+++ src/components/Navbar.jsx
@@ -6,28 +6,17 @@ import "../index.css";
 import { IconContext } from "react-icons";
 import { useRos } from "../Utils/RosConnProvider";
 import ROSLIB from "roslib";
-import TestComponent from "./TestComponent";
-import ControllerButton from "./ControllerButtons";
 
 export default function Navbar() {
-  const {ros,isCon,refresh} = useRos()
+  const { ros, isCon, refresh } = useRos();
   const [sidebar, setSidebar] = useState(false);
   const showSidebar = () => setSidebar(!sidebar);
   const location = useLocation();
   const { pathname } = location;
   const splitLocation = pathname.split("/");
-  const rosParam = new ROSLIB.Param({
-    ros: ros,
-    name: '/global_path_node/shutdown_behavior'
-  })
-  const buttonPress=()=>{
-    rosParam.get((value)=>{
-      console.log(value)
-    })
-  }
   return (
     <>
-    {/* <TestComponent></TestComponent> */}
+      {/* <TestComponent></TestComponent> */}
       <IconContext.Provider value={{ color: "undefined" }}>
         <div className="navbar">
           <Link to="#" className="menu-bar">
@@ -78,18 +67,6 @@ export default function Navbar() {
               </Link>
             </li>
           </ul>
-          <div className="controller">
-            {/* <button className="ctrl-button" >
-              <Io5.IoPlay onClick={buttonPress}/>
-            </button> */}
-            <ControllerButton/>
-            <button className="ctrl-button">
-              <Io5.IoPause />
-            </button>
-            <button className="ctrl-button">
-              <Io5.IoStop />
-            </button>
-          </div>
         </nav>
       </IconContext.Provider>
     </>
