+++ src/routes/Sensor.jsx
@@ -9,15 +9,19 @@ export default function Sensor() {
 
   return (
     <div id="sensor-tab" className="body">
-      {isCon ? <Camera/> : <Card />}
-      {isCon ? <PointCloud/> : <div className="card">
-        <h3 className="card-title"> Lidar </h3>
-        <img
-          className="sample-img"
-          src="https://s3-prod.autonews.com/s3fs-public/Velodyne_Lidar_Alpha_Prime-01_i.jpg"
-          alt="lidar"
-        />
-      </div>}
+      {isCon ? <Camera /> : <Card />}
+      {isCon ? (
+        <PointCloud />
+      ) : (
+        <div className="card">
+          <h3 className="card-title"> Lidar </h3>
+          <img
+            className="sample-img"
+            src="https://cdn.discordapp.com/attachments/1281269784415309956/1301304999296503909/image.png?ex=6723fe38&is=6722acb8&hm=2581bbd0e2f7da664dd50e680d473d7216046b9fb4de029836d4b9d8f809669c&"
+            alt="lidar"
+          />
+        </div>
+      )}
 
       <div className="card">
         <h3 className="card-title"> Sensor Fusion </h3>
